test: strengthen coverage regression assertions
The coverage tests exercised cleanup and capability branches, but two assertions did not prove the intended behavior. A failed FUSE connection could retain entity lifecycle state without failing the test, and an unavailable paginated response did not distinguish the paginated request from the legacy request.

Assert that every seeded entity lifecycle map is cleared and that pagination sends an options object. Keep the FUSE operation wiring testable through an internal module without adding those helpers to the package's public entry point.
